refactor controls, turn fly-controls into a class to simplify
