+++ app/js/controls/keys.js
@@ -1,0 +1,36 @@
+export default {
+  SHIFT: 16,
+  LEFT: 37,
+  UP: 38,
+  RIGHT: 39,
+  DOWN: 40,
+  A: 65,
+  D: 68,
+  E: 69,
+  F: 70,
+  Q: 81,
+  R: 82
+}
+
+export const xbox = {
+  left_stick_v: 1,
+  left_stick_h: 0,
+  right_stick_v: 0,
+  right_stick_h: 0,
+  A: 0,
+  B: 1,
+  X: 2,
+  Y: 3,
+  left_bumper: 4,
+  right_bumper: 5,
+  left_trigger: 6,
+  right_trigger: 7,
+  back: 8,
+  start: 9,
+  left_click: 10,
+  right_click: 11,
+  dpad_up: 12,
+  dpad_down: 13,
+  dpad_left: 14,
+  dpad_right: 15
+}
